Move errorMessage helper into modules/errors.ts
Co-locates it next to the other error-handling utilities and drops the
one-function utils/error.ts. Tree-shaking keeps it out of the IIFE
bootstrap chunk (byte-identical before/after).

Also converts repo-issue-list's `e.message` fallback to errorMessage,
and fixes a stray `:,` in the branch-priority toast noticed in review.
