+++ web_src/js/components/RepoContributors.vue
@@ -24,7 +24,7 @@ import {
   fillEmptyStartDaysWithZeroes,
 } from '../utils/time.ts';
 import {chartJsColors} from '../utils/color.ts';
-import {errorMessage} from '../utils/error.ts';
+import {errorMessage} from '../modules/errors.ts';
 import {sleep} from '../utils.ts';
 import 'chartjs-adapter-dayjs-4/dist/chartjs-adapter-dayjs-4.esm';
 import {fomanticQuery} from '../modules/fomantic/base.ts';
