+++ web_src/js/features/repo-issue-list.ts
@@ -2,6 +2,7 @@ import {updateIssuesMeta} from './repo-common.ts';
 import {toggleElem, queryElems, isElemVisible} from '../utils/dom.ts';
 import {html, htmlRaw} from '../utils/html.ts';
 import {confirmModal} from './comp/ConfirmModal.ts';
+import {errorMessage} from '../modules/errors.ts';
 import {showErrorToast} from '../modules/toast.ts';
 import {createSortable} from '../modules/sortable.ts';
 import {DELETE, POST} from '../modules/fetch.ts';
@@ -87,8 +88,8 @@ function initRepoIssueListCheckboxes() {
         await updateIssuesMeta(url, action, issueIDs, elementId);
         window.location.reload();
       } catch (err) {
-        const e = err as {responseJSON?: {error: string}, message: string};
-        showErrorToast(e.responseJSON?.error ?? e.message);
+        const e = err as {responseJSON?: {error: string}};
+        showErrorToast(e.responseJSON?.error ?? errorMessage(err));
       }
     },
   ));
