--- web_src/js/utils/error.ts
@@ -1,3 +1,0 @@
-export function errorMessage(err: unknown): string {
-  return (err as Error)?.message || String(err);
-}
