+++ web_src/js/utils/match.ts
@@ -2,7 +2,7 @@ import emojis from '../../../assets/emoji.json' with {type: 'json'};
 import {GET} from '../modules/fetch.ts';
 import {showErrorToast} from '../modules/toast.ts';
 import {parseIssuePageInfo} from '../utils.ts';
-import {errorMessage} from './error.ts';
+import {errorMessage} from '../modules/errors.ts';
 import type {Issue, Mention} from '../types.ts';
 
 const maxMatches = 6;
